23 6 Steps to Authorization Mastery

+++ app/Http/Controllers/JobController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
@@ -49,7 +52,8 @@ class JobController extends Controller
 
     public function update(Job $job)
     {
-        // authorize (On hold ... )
+        // authorize
+        Gate::authorize('edit-job', $job);
 
         // validate
         request()->validate([
@@ -70,7 +74,8 @@ class JobController extends Controller
 
     public function destroy(Job $job)
     {
-        // authorize (On hold ...)
+        // authorize
+        Gate::authorize('edit-job', $job);
 
         // delete the job
         $job->delete();
